MODULES/informationGathering/requirements.txt
beautifulsoup4
requests
python-whois
validators
